Add .gitignore, enhance build scripts, and implement relationship pathfinding feature
- Added .gitignore to exclude unnecessary files and directories.
- Updated build.ps1 and build.sh to create and activate virtual environments, install dependencies, and generate icons.
- Enhanced gedcom-dna-finder-gui.py with a new feature to find relationships between individuals.
- Introduced generate_icns.sh for creating macOS icons and modified generate_icon.py to accept input files.
- Updated gedcom-dna-finder-gui.spec to include the new icon file.

+++ build.sh
@@ -1,13 +1,47 @@
 #!/bin/bash
-rm -r dist/
-pyinstaller --noconfirm ./gedcom-dna-finder-cli.spec || { echo 'pyinstaller failed to build CLI.'; exit 1; }
-pyinstaller --noconfirm ./gedcom-dna-finder-gui.spec || { echo 'pyinstaller failed to build GUI.'; exit 1; }
-[ -d "dist" ] || { echo 'Cannot find dist build folder.'; exit 1; }
-pushd dist
-if [[ $(uname) == "Linux" ]]
-then 
-zip -r - . > ../gedcom-dna-finder-linux.zip
+[[ -e dist/ ]] && rm -r dist/
+if [[ $(uname) == "Linux" ]]; then
+	echo 'Building for Linux...'
+	out = gedcom-dna-finder-linux.zip
+  python .\generate_icon.py family_tree.ico || {
+    echo 'Failed to generate ICO file.'
+    exit 1
+  }
 else
-zip -r - . > ../gedcom-dna-finder-mac.zip 
+	echo 'Building for macOS...'
+	out = gedcom-dna-finder-mac.zip
+	./generate_icns.sh -s family_tree.png || {
+		echo 'Failed to generate ICNS file.'
+		exit 1
+	}
 fi
+[[ -e .venv/bin/activate ]] || {
+	echo 'Creating virtual environment...'
+	python3 -m venv .venv || {
+		echo 'Failed to create virtual environment.'
+		exit 1
+	}
+}
+source .venv/bin/activate || {
+	echo 'Failed to activate virtual environment.'
+	exit 1
+}
+pip install -r requirements.txt || {
+  echo 'Failed to install dependencies.'
+  exit 1
+}
+pyinstaller --noconfirm ./gedcom-dna-finder-cli.spec || {
+	echo 'pyinstaller failed to build CLI.'
+	exit 1
+}
+pyinstaller --noconfirm ./gedcom-dna-finder-gui.spec || {
+	echo 'pyinstaller failed to build GUI.'
+	exit 1
+}
+[ -d "dist" ] || {
+	echo 'Cannot find dist build folder.'
+	exit 1
+}
+pushd dist
+zip -r - . >"${out}"
 popd
